requirements.txt
openpyxl==3.0.3
Pillow==8.3.2